requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.27.0
jinja2==3.0.1